chore(ui): move logic to external file
