+++ ui/src/components/flows/MultiPanelEditorView.vue
@@ -21,7 +21,7 @@
 </template>
 
 <script setup lang="ts">
-    import {computed, onMounted, Ref, watch, ref, provide} from "vue";
+    import {computed, onMounted, Ref, watch} from "vue";
     import {useStorage} from "@vueuse/core";
     import {useStore} from "vuex";
     import {useI18n} from "vue-i18n";
@@ -30,33 +30,10 @@
     import EditorButtonsWrapper from "../inputs/EditorButtonsWrapper.vue";
     import {DEFAULT_ACTIVE_TABS, EDITOR_ELEMENTS} from "./panelDefinition";
     import {useCodePanels, useInitialCodeTabs} from "./useCodePanels";
+    import {useCodeTopology} from "./useCodeTopology";
+
     import {setupInitialNoCodeTab, setupInitialNoCodeTabIfExists, useNoCodePanels} from "./useNoCodePanels";
 
-    import {TOPOLOGY_CLICK_INJECTION_KEY} from "../code/injectionKeys";
-    import {TopologyClickParams} from "../code/utils/types";
-
-    const topologyClick = ref<TopologyClickParams | undefined>(undefined);
-    provide(TOPOLOGY_CLICK_INJECTION_KEY, topologyClick);
-
-    function findTopologyIndexes(arr: { tabs: { value: string }[] }[]): { panelIndex: number, tabIndex: number } {
-        const panelIndex = arr.findIndex(p => p.tabs.some(t => t.value === "topology"));
-        const tabIndex = panelIndex !== -1 ? arr[panelIndex].tabs.findIndex(t => t.value === "topology") : 0;
-        return {panelIndex: panelIndex !== -1 ? panelIndex : 0, tabIndex};
-    }
-
-    watch(topologyClick, (value: TopologyClickParams | undefined) => {
-        if (!value) return;
-
-        const visible = panels.value?.map((p: Panel) => p.tabs.map((t: Tab) => t.value)).flat();
-        if(!visible.includes("code") && !visible.includes("nocode")){
-            const {action, params: {section, id}} = value;
-
-            const target = findTopologyIndexes(panels.value);
-
-            if (action === "create") openAddTaskTab(target, section, id);
-            else if (action === "edit") openEditTaskTab(target, section, id);
-        }
-    })
 
     function isFlowRelated(element: Tab){
         return ["code", "nocode", "topology"].includes(element.value)
@@ -176,6 +153,8 @@
     const openTabs = computed(() => panels.value.flatMap(p => p.tabs.map(t => t.value)))
 
     const {onRemoveTab, isFlowDirty} = useCodePanels(panels)
+
+    useCodeTopology(panels, openAddTaskTab, openEditTaskTab)
 
     watch(isFlowDirty, (dirty) => {
         for(const panel of panels.value){
